backend/requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.33
flask_restful==0.3.9
python-dotenv==1.0.0
Flask-Caching==2.0.2
Flask-CORS==3.0.10
redis==4.0.2